FinalProject/lambdaLiveness/requirements_pr.txt
tensorflow==2.17.0
tensorflow-datasets==4.9.7
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.13.1
tensorflow-probability==0.24.0
pillow==11.0.0
numpy==1.26.4
glob2==0.7
keras==3.4.1